+++ src/DesktopVirtualization/DesktopVirtualization/ChangeLog.md
@@ -19,7 +19,7 @@
 -->
 ## Upcoming Release
 * Added MultiplePersistent enum value to HostPool property loadBalancerType.
-* Fixed issue when `New-AzWvdRegistrationInfo` cmdlet execution failed when public network access was disabled on the host pool.
+* Fixed issue when `New-AzWvdRegistrationInfo` cmdlet execution failed if public network access was disabled on the host pool.
 
 ## Version 5.4.7
 * Added preannouncement for breaking changes in Az.DesktopVirtualization 6.0.0